# recode/sandbox-1.0.0.145/website/docs/public/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "description": "\n- Browser\n    - CDP: [/cdp/json/version](/cdp/json/version)\n- Jupyter\n    - Notebook: [/jupyter](/jupyter)\n- MCP\n    - Streamable HTTP: [/mcp](/mcp) or [/v1/mcp](/v1/mcp)\n",
    "version": "1.0.0.143"
  },
  "paths": {
...
    "/v1/shell/exec": {
      "post": {
        "tags": [
          "shell"
        ],
        "summary": "Exec Command",
        "description": "Execute command in the specified shell session\nSupports SSE streaming if Accept header contains 'text/event-stream'",
        "operationId": "exec_command_v1_shell_exec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShellExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_ShellCommandResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "shell",
        "x-fern-sdk-method-name": "exec_command"
      }
    },
...
    "/v1/file/read": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Read File",
        "description": "Read file content",
        "operationId": "read_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_FileReadResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "file",
        "x-fern-sdk-method-name": "read_file"
      }
    },
    "/v1/file/write": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Write File",
        "description": "Write file content (supports both text and binary files)\n\nFor binary files, set encoding to 'base64' and provide base64-encoded content.\nFor text files, use default 'utf-8' encoding.",
        "operationId": "write_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_FileWriteResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "file",
        "x-fern-sdk-method-name": "write_file"
      }
    },
...
    "/v1/file/list": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "List Path",
        "description": "List path contents with flexible options",
        "operationId": "list_path_v1_file_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_FileListResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "file",
        "x-fern-sdk-method-name": "list_path"
      }
    },
...
    "/v1/jupyter/execute": {
      "post": {
        "tags": [
          "jupyter"
        ],
        "summary": "Execute Jupyter Code",
        "description": "Execute Python code using Jupyter kernel with session persistence\n\nThis endpoint allows you to execute Python code and get results back.\nYou can optionally specify a kernel_name (defaults to 'python3').\nUse session_id to maintain variable state across multiple requests.\nSessions automatically expire after 30 minutes of inactivity.",
        "operationId": "jupyter_execute_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JupyterExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_JupyterExecuteResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-method-name": "execute_code",
        "x-fern-sdk-group-name": "jupyter"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BashCommandStatus": {
        "type": "string",
        "enum": [
          "running",
          "completed",
          "no_change_timeout",
          "hard_timeout",
          "terminated"
        ],
        "title": "BashCommandStatus",
        "description": "Shell command execution status (compatible with OpenHands)"
      },
...
      "ConsoleRecord": {
        "properties": {
          "ps1": {
            "type": "string",
            "title": "Ps1",
            "description": "Command prompt"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Executed command"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "description": "Command output",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "ps1",
          "command"
        ],
        "title": "ConsoleRecord",
        "description": "Shell command console record model"
      },
...
      "FileContentEncoding": {
        "type": "string",
        "enum": [
          "utf-8",
          "base64",
          "raw"
        ],
        "title": "FileContentEncoding",
        "description": "File content encoding type"
      },
...
      "FileInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "File name"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Full file path"
          },
          "is_directory": {
            "type": "boolean",
            "title": "Is Directory",
            "description": "Whether it's a directory"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "File size in bytes"
          },
          "modified_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified Time",
            "description": "Last modified time (ISO format)"
          },
          "permissions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permissions",
            "description": "File permissions"
          },
          "extension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension",
            "description": "File extension"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "is_directory"
        ],
        "title": "FileInfo",
        "description": "File information"
      },
      "FileListRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Directory path to list"
          },
          "recursive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recursive",
            "description": "Whether to list recursively",
            "default": false
          },
          "show_hidden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Hidden",
            "description": "Whether to show hidden files",
            "default": true
          },
          "file_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Types",
            "description": "Filter by file extensions (e.g., ['.py', '.txt'])"
          },
          "max_depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Depth",
            "description": "Maximum depth for recursive listing"
          },
          "include_size": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Size",
            "description": "Whether to include file size information",
            "default": true
          },
          "include_permissions": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Permissions",
            "description": "Whether to include file permissions",
            "default": false
          },
          "sort_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By",
            "description": "Sort by: name, size, modified, type",
            "default": "name"
          },
          "sort_desc": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Desc",
            "description": "Sort in descending order",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "FileListRequest",
        "description": "File list request"
      },
      "FileListResult": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Listed directory path"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileInfo"
            },
            "type": "array",
            "title": "Files",
            "description": "List of files and directories",
            "default": []
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items",
            "default": 0
          },
          "directory_count": {
            "type": "integer",
            "title": "Directory Count",
            "description": "Number of directories",
            "default": 0
          },
          "file_count": {
            "type": "integer",
            "title": "File Count",
            "description": "Number of files",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "FileListResult",
        "description": "File list result"
      },
      "FileReadRequest": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Absolute file path"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line",
            "description": "Start line (0-based)"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line",
            "description": "End line (not inclusive)"
          },
          "sudo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sudo",
            "description": "Whether to use sudo privileges",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "FileReadRequest",
        "description": "File read request"
      },
      "FileReadResult": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "File content"
          },
          "file": {
            "type": "string",
            "title": "File",
            "description": "Path of the read file"
          }
        },
        "type": "object",
        "required": [
          "content",
          "file"
        ],
        "title": "FileReadResult",
        "description": "File read result"
      },
...
      "FileWriteRequest": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Absolute file path"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Content to write (text or base64 encoded for binary)"
          },
          "encoding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileContentEncoding"
              },
              {
                "type": "null"
              }
            ],
            "description": "Content encoding: utf-8 for text, base64 for binary data",
            "default": "utf-8"
          },
          "append": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Append",
            "description": "Whether to use append mode",
            "default": false
          },
          "leading_newline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leading Newline",
            "description": "Whether to add leading newline (only for text mode)",
            "default": false
          },
          "trailing_newline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Newline",
            "description": "Whether to add trailing newline (only for text mode)",
            "default": false
          },
          "sudo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sudo",
            "description": "Whether to use sudo privileges",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file",
          "content"
        ],
        "title": "FileWriteRequest",
        "description": "File write request - supports both text and binary content"
      },
      "FileWriteResult": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Path of the written file"
          },
          "bytes_written": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes Written",
            "description": "Number of bytes written"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "FileWriteResult",
        "description": "File write result"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JupyterExecuteRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Python code to execute"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout",
            "description": "Execution timeout in seconds",
            "default": 30
          },
          "kernel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kernel Name",
            "description": "Kernel name to use (e.g., 'python3', 'python3.11', 'python3.12'). Defaults to 'python3'",
            "default": "python3"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Session ID to maintain kernel state across requests"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "JupyterExecuteRequest",
        "description": "Jupyter code execution request model"
      },
      "JupyterExecuteResponse": {
        "properties": {
          "kernel_name": {
            "type": "string",
            "title": "Kernel Name",
            "description": "Name of the kernel used for execution"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session ID for this kernel instance"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status: ok, error, or timeout"
          },
          "execution_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Count",
            "description": "Execution count from the kernel"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/JupyterOutput"
            },
            "type": "array",
            "title": "Outputs",
            "description": "List of execution outputs"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "The executed code"
          },
          "msg_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msg Id",
            "description": "Message ID from Jupyter kernel"
          }
        },
        "type": "object",
        "required": [
          "kernel_name",
          "session_id",
          "status",
          "outputs",
          "code"
        ],
        "title": "JupyterExecuteResponse",
        "description": "Jupyter code execution response model"
      },
...
      "Response_FileListResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileListResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[FileListResult]"
      },
      "Response_FileReadResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileReadResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[FileReadResult]"
      },
...
      "Response_FileWriteResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileWriteResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[FileWriteResult]"
      },
...
      "Response_JupyterExecuteResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JupyterExecuteResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[JupyterExecuteResponse]"
      },
...
      "Response_ShellCommandResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShellCommandResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[ShellCommandResult]"
      },
...
      "ShellCommandResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Shell session ID"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Executed command"
          },
          "status": {
            "$ref": "#/components/schemas/BashCommandStatus",
            "description": "Command execution status"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output",
            "description": "Command execution output, only has value when status is completed"
          },
          "console": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConsoleRecord"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Console",
            "description": "Console command records"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code",
            "description": "Command execution exit code, only has value when status is completed"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "command",
          "status"
        ],
        "title": "ShellCommandResult",
        "description": "Shell command execution result model"
      },
...
      "ShellExecRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique identifier of the target shell session, if not provided, one will be automatically created"
          },
          "exec_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exec Dir",
            "description": "Working directory for command execution (must use absolute path)"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Shell command to execute"
          },
          "async_mode": {
            "type": "boolean",
            "title": "Async Mode",
            "description": "Whether to execute command asynchronously (default: False for async, False for synchronous execution)",
            "default": false
          },
          "timeout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout",
            "description": "Maximum time (seconds) to wait for command completion before returning running status"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "ShellExecRequest",
        "description": "Shell command execution request model"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}